+++ src/linter/config-presets/aglint-recommended.ts
@@ -3,9 +3,12 @@
  * which are useful for all projects.
  */
 
+import { AdblockSyntax } from '@adguard/agtree';
+
 import { type LinterConfig } from '../common';
 
 const config: LinterConfig = {
+    syntax: [AdblockSyntax.Common],
     rules: {
         'duplicated-hint-platforms': 'error',
         'duplicated-hints': 'error',